app-emulation/open-vm-tools: fix install with pam disabled
Bug: https://bugs.gentoo.org/601640
Package-Manager: portage-2.3.2_p13

+++ app-emulation/open-vm-tools/open-vm-tools-10.1.0.ebuild
@@ -122,8 +122,10 @@ src_install() {
 	default
 	prune_libtool_files --modules
 
-	rm "${ED%/}"/etc/pam.d/vmtoolsd || die
-	pamd_mimic_system vmtoolsd auth account
+	if use pam; then
+		rm "${ED%/}"/etc/pam.d/vmtoolsd || die
+		pamd_mimic_system vmtoolsd auth account
+	fi
 
 	newinitd "${FILESDIR}/open-vm-tools.initd" vmware-tools
 	newconfd "${FILESDIR}/open-vm-tools.confd" vmware-tools
